refactor separate kernel and process strategy
